% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_data.R
\name{get_csv_data}
\alias{get_csv_data}
\title{Get CSV Data}
\usage{
get_csv_data(url)
}
\arguments{
\item{url}{specifies the path to the file}
}
\value{
A data.frame
}
\description{
This function retrieves and loads csv data from a particular URL.
It does not care if the file is store remotely or on you local machine.
It converts data and returns it as a data.frame
}
